+++ oil-service/oil-service-impl/src/main/java/com/oil/av/dao/platform/config/SmsQueueDao.java
@@ -44,7 +44,7 @@ public interface SmsQueueDao {
      * @param  smsQueueEntity
      * @return
      */
-	Long insert(SmsQueueEntity smsQueueEntity);
+	Integer insert(SmsQueueEntity smsQueueEntity);
 	
 	/**
      * 更新运营管理-短信队列
